% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.stat.validation.R
\docType{data}
\name{sum.stat.validation}
\alias{sum.stat.validation}
\title{UK Biobank GWAS summary statistics on height around rs11090631 on chromosome 22}
\format{1 row per SNP. The variables are as follows:
\itemize{
   \item SNP, SNP ID
   \item A1, effect allele
   \item A2, reference allele
   \item Freq1, the allele frequency of Allele1
   \item b, estimate of marginal effect in GWAS
   \item se, standard error of the estimates of marginal effects in GWAS
   \item N, sample size
}}
\source{
Height GWAS results for 120,286 genetically British individuals in UK Biobank Wave 1 data. 
The phenotypes were age and sex adjusted and then inverse-gaussian transformed.
}
\usage{
data(sum.stat.validation)
}
\description{
This data set is the GWAS summary statistics on height computed using UK Biobank Wave 1 data. 
582 SNPs within a 1-Mb centred at rs11090631 are included in the data.
}
\author{
Zheng Ning, 2018-05-29
}
